requirements.txt
feedparser
requests
beautifulsoup4
Pillow